1.1.0

+++ include/basen/hash/sha256.hpp
@@ -6,6 +6,6 @@
 
 namespace hash
 {
-	void sha256(const uint8_t *data, uint64_t data_size, uint8_t *hash) noexcept;
+	void sha256(const uint8_t *data, size_t data_size, uint8_t *hash) noexcept;
 	std::vector<uint8_t> sha256(std::span<const uint8_t> data) noexcept;
 }
